moved get_logger back
