+++ src/co_tools/co_utils.py
@@ -3,7 +3,7 @@ import subprocess
 from multiprocessing import cpu_count
 
 # local imports
-from co_tools.co_logger.get_logger import LOGGER
+from co_tools.get_logger import LOGGER
 
 co_cpus = os.getenv("CO_CPUS")
 aws_batch_job_id = os.getenv("AWS_BATCH_JOB_ID")
