+++ src/co_tools/get_dir_contents.py
@@ -2,7 +2,7 @@
 import sys
 
 # local imports
-from co_tools.co_logger.get_logger import LOGGER
+from co_tools.get_logger import LOGGER
 from co_tools.co_utils import get_dir_contents
