+++ src/co_tools/get_read_pattern.py
@@ -2,7 +2,7 @@
 import sys
 
 # local imports
-from co_tools.co_logger.get_logger import LOGGER
+from co_tools.get_logger import LOGGER
 from co_tools.co_fastq import get_read_pattern
